Add files via upload
In this update what I thought of was eliminating unnecessary portions of the images to crop and keep only the useful region.
So, cropping section has been added which will keep only the face and discard the rest. In this manner we should be minimising the
noise and achieve better results.